Support #37323 - Display location data added via workbook upload
- Removed GeographySearchDialog, this component is not used and it's a duplicate of the GeographySearchBox completely.
- Added an external link icon to the view details button for a linked geographic place.
- Extracted the geography search box to it's own component, should not be completely part of the CollectingEventForm.
