+++ packages/dina-ui/components/collection/collecting-event/GeographySearchBox.tsx
@@ -12,6 +12,7 @@ import { ReactNode, useState } from "react";
 import useSWR from "swr";
 import { DinaMessage, useDinaIntl } from "../../../intl/dina-ui-intl";
 import { TbMapPinX } from "react-icons/tb";
+import { FaArrowUpRightFromSquare } from "react-icons/fa6";
 interface GeographySearchBoxProps {
   inputValue: string;
   onInputChange: (value: string) => void;
@@ -143,7 +144,11 @@ export function GeographySearchBox({
             aria-label={formatMessage("locationLabel")}
             className="form-control"
             placeholder={formatMessage("locationPlaceholder")}
-            onChange={(e) => onInputChange((e.target as HTMLTextAreaElement | HTMLInputElement).value)}
+            onChange={(e) =>
+              onInputChange(
+                (e.target as HTMLTextAreaElement | HTMLInputElement).value
+              )
+            }
             onFocus={(e) => (e.target as HTMLInputElement).select()}
             onKeyDown={(e) => {
               if (e.keyCode === 13) {
@@ -211,6 +216,7 @@ export function GeographySearchBox({
                   rel="noopener noreferrer"
                 >
                   <DinaMessage id="viewDetailButtonLabel" />
+                  <FaArrowUpRightFromSquare style={{ marginLeft: "0.3em" }} />
                 </a>
               </div>
             </div>
